Wide area spawning turfs no longer keep old turf's material.
This often just caused a bunch of issues in some turfs randomly receiving some material while others didn't.
I don't see a need for it but if there's some potential use case that I can't come up with we could add a toggle later down the road. As is however, it's usually just a hassle and annoyance most of the time.