Adds drop_loot_on_death component.
Component, can be added to any mob or obj critter.
Accepts any number of object types to drop.
When the parent of the component dies, every type provided to the component is spawned at the parent's location along with a message informing of such. The component removes itself afterwards.
Duplicate components' contents are transfered to the original one, to avoid clutter in the drop message.