Adds "button" sprites for a full bookcase.
Some bookcases have a special sprite variant and functionality with one of the books being slightly off, you can push it in and it acts as a trigger for adventure wizard things.
I noticed that some of my bookcases weren't animating in my prefab and I assumed it was an issue with the icon_state.
Anyways, long story short, turns out the bookcases I was trying to "fix" were just decoration and never connected to things, but before that I've already made sprites for a "button" variant of the bookcase so I guess we have those now for any future uses.