Big change - add numbers to some objects for amounts / count / ammo.
This creates a new object / vis_contents when you hold an item in
your hands, showing (by default) the amount.

It's manually updated because there's no real standard for how to
change an item's amount - it'd be nice to standardize this in the
future and this can probably help point out where to make changes.

Tested most of this. Some items will bug and not update right
away or show the wrong number in some cases, but for the moooost
part it should be relatively good.