break everything forever
ok so now that thats out of the way

* change round startup in a lot of ways
* gameticker is now inited immediately instead of last
* tgui controller is also inited immediately
* this changes how some stuff works, mostly by making the process
  scheduler have a list of things that have already been created
  to add but NOT set up
* the preround countdown will now hang at 30 if setup isnt done
* map vote inits at T-60s instead when the countdown starts
* some startup sequences have more progress info on title screen
* fixed an errant <h3> in the process scheduler context thing
* more, probably

if at any time you are not satisfied with this commit,
please feel free to roll it back.
thank you.
discount coder.

anyway i checked a bunch of stuff and got no runtimes or wonky
behavior but i didn't check everything and you never know what
will show up when you have 100 people instead of 1 hammering it.