Fix Oshan hallway becoming vented at round start
The hallway by the engineering/mechanic labs had stairs, but
they were of the airless type. Somehow this managed to drain
the *entire hallway* of air fairly consistently, which seems
like a bug in and of itself (3 tiles of zero atmos should not
render a hundreds-of-tiles-long hallway of almost all air)

Soooooo, that's fixed. No more airless stairs.