Fix the King in Yellow runtiming.
This has apparently been broken since May :(

Also, tooltips caused some funkiness where the first person to even
hover over it (to generate a tooltip) would become a reader, which
isn't... quite right. So that's different, now, maybe.

I dunno. There's no way to tell easily if a description fetch is
because of a tooltip or an interaction, so this memetic thing
doesn't quite work.

...

That makes me wonder if the toolbox is broken like this, too.
ugh