Attempt to preload-by-browse_rsc() manufacturer images.
So the way this works:
- Generate icons for all datum/manufacture during round setup.
- Later, just before round setup is done, send these icons to clients.
- Any client that connects afterwards gets them upon connect.

The goal is that the icons are sent up front, so they don't need
to be sent later when you actually want to use it. Of course, this
causes a bunch of waste (everyone gets the icons even if they never
use the manufacturer) but there's no better way to do it atm.

Ideally most icons could be generated and put on a cdn or something,
or maybe byond's browse_rsc() could suck less ass, like maybe
batch-sending resources (hey, an assoc list?!), but hahahahhahahhahahahahahaha