adds aerosol reaction to golem death
this copies the contents of a golem to a temporary container on death so that both the current smoke reaction and an additional small aerosol reaction occur. the idea is making golems of disease reagents and various poisons dangerous to fight at close-range, so that more exotic golems aren't easily handled by kiting with a basic melee weapon and walking off after.