Refactor end-of-round output
Previously, every single part ended up doing a boutput(world).
Now (much like the AI change), these are combined into a single
list and then joined and output at the end.
This should help with the huge scrolling mess at round-end.

AI law output has also been changed:
Previously, the laws said "The AI (player) had the following...",
and depended on there actually being an AI. This is a long-standing
relic from when the AI had to be alive to update the laws, and
when it was only ever possible to have a single AI.

That's not the case any more, so it now says "AI and Cyborgs had
the following laws", and lists them from the centralized lawset
rather than using the AI player to do it.

A change like this probably has some issues, but I tested a quick
round-end locally and everything seemed to work fine, so... vOv