Lower screen shake intensity across the board
Screen shake proc changed: the intensity argument was multiplied
by 32 (the size of a tile); this is no longer the case. It is now
used straight as a per-pixel amount to move the screen (max)

Accordingly every call to shake_screen has been changed,
generally multiplying the intensity by 4~8 (depending on case).
I'll admit it's very arbitrary how things were changed, but
the goal was to lower it to values that would not be so high
as to make people quit the game just to avoid it.

In some cases the original intensity was so outrageous, e.g.
manta's start-or-stop, that the screen sould move up to 5 whole
tiles in any direction. Consider that the view is 15 tiles tall
and you're possibly throwing a full third of the screen
out of view!

Some effects will likely need to be adjusted after this.