Allow bolt+toggle to occur in the same click
This allows AIs and cyborgs to bolt doors open in a single click, which
can be very helpful with certain Ion Storm laws.

Also fixes a bug where only one of the two main power wires and only one
of the two backup power wires in the door would respond to the
multitool.