Remove the "human" requirement from law 2.
With the way the law was previously written, a lizard or cow head of
staff would not be included in the chain of command, which was extremely
tedious to deal with for the AI.

This does not modify law 1 in any way. The interaction between the AI
and anti-griefing rules has not been touched.

This does affect non-human antagonists. Now, if a changeling asks the AI
to do something, they have to be overridden by someone higher-ranking,
like a staff assistant or clown. Changelings who have eaten the captain
are not the captain, after all.

This matches the previous interaction between law 2 and nuclear
operatives. As they are not in the station's chain of command, anyone
who is part of the crew can override their orders.