Allow PDA messages to be sent to multiple groups at the same time.
This includes messages that are not broadcast messages (i.e. messages
to a single PDA).

To receive a message sent to multiple groups, you need to be in at least
one of the groups and be muting at most zero of the groups it is
addressed to.

Change MAILBOT messages to send to multiple groups, including a new
"alerts" group for each message type. PDAs have an option to mute the
alerts groups that are relevant to them.

Remove the special properties of the "ai" mailgroup. Instead, add the AI
to all default groups.

Messages sent to multiple groups cannot be group-replied to by clicking
the group name.

Add genetics to the medbay group, muting it by default.

Mute genebooth sales alerts for roboticists by default.