Rework PDA alert routing (#3178)
* add constants for hard-coded mail group names

* Allow PDA messages to be sent to multiple groups at the same time.

This includes messages that are not broadcast messages (i.e. messages
to a single PDA).

To receive a message sent to multiple groups, you need to be in at least
one of the groups and be muting at most zero of the groups it is
addressed to.

Change MAILBOT messages to send to multiple groups, including a new
"alerts" group for each message type. PDAs have an option to mute the
alerts groups that are relevant to them.

Remove the special properties of the "ai" mailgroup. Instead, add the AI
to all default groups.

Messages sent to multiple groups cannot be group-replied to by clicking
the group name.

Add genetics to the medbay group, muting it by default.

Mute genebooth sales alerts for roboticists by default.

* Mute all Cargo-related alert groups (including mailalerts) for the AI by default.

Fix CYBORG-DAEMON's messages being missed in the *-MAILBOT changes.

Add the silicon group to the CYBORG-DAEMON mailing list.

* Add some PDA alerts I missed.

Fix group filtering not affecting single-target messages due to indentation.

* easter egg: experimental equality law module creates the silicon department on the messaging app