Add shoving crayons up others' noses
Crayons can now be shoved up the noses of carbon-based life.
Disgusting, I know. This is very vital to Clown and Mime gameplay.
It's a wonder this was not added sooner!

When shoved into someone's nose, a crayon will deal 15 BRAIN, and
add 3 units of lithium to the bloodstream of the victim, to induce
drooling and wonky movement.

In order to inflict severe brain damage, you must shove 4 crayons
into someone's nose. If you want to kill someone you must shove
an entire box of crayons, plus an extra finishing crayon (a total
of 8) in order to kill them.

Have fun!