implant: allow implanting individual limbs in the body
Add the ability to implant and remove implants from individual
limbs, for implants to specify if they should create paper
wads when removed (i.e. bullets, crayons, etc.), and for
implanters to specify a mandatory body part to implant to.

This change is neccessary to implement the crayon implant, which
must live in the head. This change has been requested by multiple
people, does not have any impact on the current surgery tree,
and implants will default to the chest if they do not specify what
body part they should be placed in.

In specific, the following implanters have locked body parts:
  - /obj/item/implanter/microbomb/ can only implant in the chest.
  - /obj/item/implanter/macrobomb/ can only implant in the chest.
  - /obj/item/implanter/uplink_macrobomb can only implant in the
    chest.
  - /obj/item/implanter/uplink_microbomb can only implant in the
    chest.
  - /obj/item/implanter/mindslave can only implant in the head.
  - /obj/item/implanter/super_mindslave can only implant in the
    head.

Upon the severance of a limb, the implants that were in it are
transferred to the limb, and are removed from the main mob's
implant list, and the implant's on_death() proc is called.
Upon the reattachment of a limb, the implants in the limb are
implanted into the mob receiving the limb, and the implanted()
proc is called, with the attacher being the implanter.