crayons: implant crayon into head instead of adding lithium
Change the behaviour of the crayon to implant a crayon into the
head of the target instead of adding lithium to their bloodstream.
This implant has a 20% chance of dealing 1 BRAIN per life cycle,
and random emotes. There is a 1% chance of the implant owner
sneezing up the crayon.

When the crayon is shoved up someone's nose, the fingerprints
of the attacker is applied to the crayon to allow for forensic
analysis. When the implant is removed, the crayon originally
shoved takes its place, and is also stained with the blood of
whomever it was in.