Multi-Pod Cloning! (#3163)
* Multi-Pod Cloning!

Allows cloning consoles to connect to up to 3 cloning pods at once.
Cloning pods must be within 4 tiles of the console (same as before).

Gene Power Module, the (currently unused) Mind Wipe Module, and the
genetics setting have been moved from the pod to the console. They were
already added and removed via the console, so this only affects
scenarios where one part of the cloning system is completely destroyed
and needs to be replaced.

Pressing the clone button will use the cloning pod with the highest
amount of stored material, avoiding cloning pods that are currently
working on a clone. (Pods preparing a body are fine, however.) If that
pod is unsuitable for a different reason, tough.

Moves cloning pod status to its own tab of the cloning console UI. The
names of pods can be changed using a hand labeler.

The Syndicate UI theme will be used if *any* of the attached pods have a
mindslave module attached.

* update maps that explicitly set the genetics research setting on a cloning pod to set it on the console instead

* fix merge error

Co-authored-by: pali <6pali6@gmail.com>