feature: add a syringe gun that accepts love & hugs - the love gun
No more tomorrow, the time is today.  It can make you feel okay.
You pull the trigger of the Love Gun.

this adds the "love gun" as an alternative to the syringe gun.
the love gun accepts only two reagents (love & hugs) and cannot
be emagged to accept others.  it's for love, not war.
because of this limitation, it has a larger capacity than the
standard syringe gun

this also adds the "Love Gun Plus", for when the station is really
in need of more love than the Love Gun carries.  this was mostly added
to be used in Lily White's office, but I bet it'd be neat to stash
other harder to reach areas