Feature: Add the Love Gun (#4543)
* feature: add a syringe gun that accepts love & hugs - the love gun

No more tomorrow, the time is today.  It can make you feel okay.
You pull the trigger of the Love Gun.

this adds the "love gun" as an alternative to the syringe gun.
the love gun accepts only two reagents (love & hugs) and cannot
be emagged to accept others.  it's for love, not war.
because of this limitation, it has a larger capacity than the
standard syringe gun

this also adds the "Love Gun Plus", for when the station is really
in need of more love than the Love Gun carries.  this was mostly added
to be used in Lily White's office, but I bet it'd be neat to stash
other harder to reach areas

* chore: use the love gun plus in rodney's office

confirmed with rodney on discord that this is lovely as
long as it has a whole lotta love in it

* chore: add the love gun to the clownspider prefabs

this includes the love gun with the pink clown clothes in the
clownspider prefabs for both the debris field & the trench