feature: apply sonic stun from flashbang to deaf mobs (#4564)
if the mob is deaf and is not wearing the auditory headset
we should still apply the sonic stun to them.

this does not affect users wearing earplugs or earmuffs because of space magic

balance: it's become a bit of a "thing" on the main servers to run with the
deaf trait. it's a +1 trait point.  however, if you pair it with sechuds, when you
go up against any security force you pop out the auditory headset for complete
immunity to flashbangs.  this same immunity applies when going up against vampires
or shamblers.