Adds client image groups.
Could be used for handling groups of mobs and images that those mobs should see. Specific use example: group of mobs wearing prodoc health goggles and images of health icons that they should see.
Added or removed images are added and removed from all of the mobs in the group.
Added or removed mobs gain or lose visibility of the images in the group.