fix: use turf for nearby sound calculation in playsound
because we're expecting to work with the object itself, we can't always `GET_NEARBY` an item someone is holding or an item that's in locker-space, etc.  their Z-level is not valid for the nearby lookup.

instead, we need to `GET_NEARBY` the `source_turf` which is more likely to be placed and have a `Z` associated with it

fixes #4869