Ghost VR / Afterlife Bar mobs're now included in random antag events
This should make it so that players using either of those
areas will be available for random event antagonist events.

(I tested this in several cases and it works as expected, so
living people in VR cannot get chosen and bar visitors shouldn't
be able to carry anything with them back to the living)

Please report any weirdness!