Check the component's other props too (read desc)
After adding the onComponentShouldUpdate hook, ReagentInfo components
no longer updated when props other than the container prop changed.

This was a known issue, so to address concerns about components not updating when expected,
I've fixed it to also update if any non-container props change,
using a shallow difference check.

Maybe it would be worth it to
add a common function for this sort of purpose later,
like "primitiveShallowDiffers(a, b)" that only checks non-object values,
or like "exclusiveShallowDiffers(a, b, excludedProps...)"
that lets you specify properties to exclude from the check.