Adds in a new chaplain antagonist item <!-- The text between the arrows are comments - they will not be visible on your PR. --> <!-- To automatically tag this PR, add the uppercase label(s) surrounded by brackets below, for example: [LABEL] --> [INPUT WANTED] ## About the PR <!-- Describe the Pull Request here. What does it change? What other things could this impact? --> This PR adds in a new chaplain antag item, a scroll that once read, replaces a chaplains heart and grants them the ability to summon a slightly less powerful syndicate knife for 8 seconds with a cooldown of 30 seconds
## Why's this needed? <!-- Describe why you think this should be added to the game. -->
The chaplain has very little in terms of unique antag items, and syndicate knives are fun to use
The name's could use some work however

## Changelog
<!-- If necessary, put your changelog entry below. Otherwise, please delete it.
Use however you want to be credited in the changelog in place of CodeDude.
Use (*) for major changes and (+) for minor changes. For example: -->
```changelog
(u)Carsontheking
(+)Adds In a new antag item for chaplains!
```