If someone has the arrest status while going through the brig chute, set their record to "released"
<!-- The text between the arrows are comments - they won't be visible on your PR. -->
<!-- To automatically tag this PR, add the label(s) surrounded by brackets anywhere, for example: [LABEL] -->
<!-- PRs should at least have one area (A-) label and at least one category (C-) label -->

## About the PR <!-- Describe the Pull Request here. What does it change? What other things could this impact? -->

## Why's this needed? <!-- Describe why you think this should be added to the game. -->

## Changelog
<!-- If necessary, put your changelog entry below. Otherwise, please delete this section.
Use however you want to be credited in the changelog in place of CodeDude.
Use (*) for major changes and (+) for minor changes. For example: -->

```changelog
(u)CodeDude
(*)Added soft soft pizza to the game - a tasty drink found in soda machines!
(+)Made some small sprite changes to the base pizza sprite.
```