General changes to rattlesnake venom and behavior [LABEL][AI][Chemistry][Balance] ## About the PR <!-- Describe the Pull Request here. What does it change? What other things could this impact? --> This PR: Makes it so rattlesnakes no longer bite borgs and chase after people if their personal space is intruded. Changes the viper venom stagger effect for a 4 second stun. Removes the random +8 tox chance. Adds a small amount of bleed to complement the more dynamic take_bleeding_damage proc.
## Why's this needed? <!-- Describe why you think this should be added to the game. -->
With the advent of the new critter spawn event, i've noticed that encounters with rattlesnakes tend to be a lot more aggressive on the part of the snake than i'd like. This should make them more forgivable to be around.
As for the venom changes, the 10% chance of a +8 tox makes it wildly inconsistent, and stagger is rarely used in chemicals due to being mostly a combat effect, so i've changed those. As for the bleeding, a bump up on the bleeding speed should compensate for them no longer continuously biting you + the removal of the random extra damage.

## Changelog <!-- If necessary, put your changelog entry below. Otherwise, /please/ delete this entire section. -->

```changelog
(u)Hans
(+)Rattlesnakes will no longer chase people down unless directly attacked, they will still bite you if you get close.
```