emergency shutoff switch take 2, now with 3 minutes of waiting
previous solution had a big issue, namely that the apiError proc
is only called when the request's retries have been exhausted

which is a huge wait (at least 2:30 before the requests themselves),
so it wasn't really useful at stopping a flood.

now it does it before it retries, and logs the current fail count.