rm computerx
If this was ever to be worked on in earnest again, it'd probably be a whole lot easier to just use TGUI and flex/grid systems provided that way. There'd also be a lot less silly winset code to get the window to update dynamically, when you can just push the new data. It's also 11 years old.