there's a funny story behind this change
the short version of it is that, before i was a coder,
i asked for someone to add the current map to the logs.
that way, i could potentially add in something that would
show where on the map someone said/did something.

someone added it... to the actual world.log output,
which isn't output in the full HTML logs that
the log viewer actually uses. oops.

anyway here we are years later. hi