Adds renderButtonsDeps prop to ReagentList
Here's the problem: If you pass an anonymous function to renderButtons, then it's going to change every time the parent component of ReagentList renders. If the parent component uses useBackend, it's going to re-render almost every second or so. Therefore the ReagentList will also re-render with it, because the renderButtons prop keeps changing.

To fix this, renderButtons will no longer trigger a re-render when it changes. Instead, the new renderButtonsDeps prop, which represents the dependencies of the renderButtons function, will trigger a re-render when it is changed. So if any of the dependencies of renderButtons change, then the renderButtonsDeps prop will change, and THAT will cause an actually necessary re-render.