Update kondaru.dmm
<!-- The text between the arrows are comments - they won't be visible on your PR. -->
<!-- To label this PR, add the label(s) without the prefixes surrounded by brackets anywhere, for example: [LABEL] -->
<!-- PRs should at least have one area (A-) label and at least one category (C-) label -->

## About the PR <!-- Describe the Pull Request here. What does it change? What other things could this impact? -->
This just makes a few changes to Kondaru's toxins to make it more in line with other maps' toxins!


## Why's this needed? <!-- Describe why you think this should be added to the game. -->
Currently, toxins is almost unplayable on Kondaru. It's a pain in the ass to do anything with just one burn chamber, plus the layout is a bit cramped, and generally it's just not the best to look at. 

## Changes
Moved things around to make room for a second burn chamber
Im gonna be honest, this is a pretty close ripoff of donut 3's toxins, but why mess with greatness? 
Added some lab tables, cause those look cool 

## Before and After!
Before:
![image_2024-07-03_123035276](https://github.com/goonstation/goonstation/assets/170150086/1bf42c3c-fee4-4860-a552-22a74c9cada3)
After:
![image_2024-07-03_123204661](https://github.com/goonstation/goonstation/assets/170150086/d9765375-f447-455a-a74e-a15f3eb0732e)