Removes all improper datum[] uses
This is undefined behavior which is being made to error in 516. Best to
yeet it.

Most of these are swaps to .vars or just .
There is one set of swaps to : because the typing seems... ambigious and
I'm not sure what's meant to be going on with it. If you have opinions
lemme know and I can try strictly typing it.