Adds welded sheet masks as a craftable item [LABEL][feature][clothing][sprites] <!-- PRs should at least have one area (A-) label and at least one category (C-) label -->
## About the PR <!-- Describe the Pull Request here. What does it change? What other things could this impact? -->

This PR adds the steel sheet welded mask. It posseses good head protection and some disorient resist from flashes, at the cost of lowering visibility. It can be crafted by using a welder on a stack of at least 5 steel sheets, and can be altered with a welder to increase visibility, at the cost of halving protection and removing most disorient resist.

## Why's this needed? <!-- Describe why you think this should be added to the game. -->

I've been wanting some different mask options for quite a bit, and this felt like a nice option with some usability, but a mostly cosmetic alternative version as well.

## Changelog

```changelog
(u)Colossus
(+)Added metal masks to the game, which can be crafted by using a welder on a stack of 5 or more steel sheets. Masks offer protection in exchange for visibility.
```